requirements.txt
annotated-types==0.7.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dysts==0.96
fonttools==4.61.0
future==1.0.0
gluonts==0.16.2
jax==0.8.1
jaxlib==0.8.1
jaxopt==0.8.5
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
llvmlite==0.45.1
MarkupSafe==3.0.3
matplotlib==3.10.7
ml_dtypes==0.5.4
nolds==0.6.3
numba==0.62.1
numpy==2.1.3
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
threadpoolctl==3.6.0
toolz==0.12.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
